gaep_learn/requirements.txt
Flask==0.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10
itsdangerous==0.24
google-api-python-client==1.6.2
GoogleAppEngineCloudStorageClient==1.9.22.1